Full authorization completed

+++ fantasy-beans/src/main/java/repos/ElfRepositoryBean.java
@@ -1,7 +1,12 @@
 package repos;
 
 import entities.Elf;
+import entities.Forest;
+import entities.Role;
+import entities.User;
 
+import javax.ejb.EJB;
+import javax.ejb.EJBException;
 import javax.ejb.Remote;
 import javax.ejb.Stateless;
 import javax.persistence.EntityManager;
@@ -17,8 +22,15 @@ public class ElfRepositoryBean implements ElfRepository {
   @PersistenceContext(unitName = "postgresDb")
   EntityManager entityManager;
 
+  @EJB
+  private UserRepository userRepository;
+
+  @EJB
+  private ForestRepository forestRepository;
+
   @Override
-  public void save(Elf elf) {
+  public void save(Elf elf, int fromId) {
+    authorize(fromId, elf.getForest().getId());
     entityManager.persist(elf);
   }
 
@@ -38,13 +50,24 @@ public class ElfRepositoryBean implements ElfRepository {
   }
 
   @Override
-  public void removeById(int id) {
+  public void removeById(int id, int fromId) {
     Elf elf = getById(id);
+    authorize(fromId, elf.getForest().getId());
     entityManager.remove(elf);
   }
 
   @Override
-  public void update(Elf elf) {
+  public void update(Elf elf, int fromId) {
+    authorize(fromId, elf.getForest().getId());
     entityManager.merge(elf);
+  }
+
+  private void authorize(int fromId, int forestId) {
+    Forest forest = forestRepository.getById(forestId);
+    User user = userRepository.getById(fromId);
+
+    if (!user.getRole().equals(Role.ADMIN) && !(fromId == forest.getUser().getId())) {
+      throw new EJBException("Not authorized");
+    }
   }
 }
